Extract fallback mechanism.
