+++ fallback/src/main/java/io/knotx/fallback/FallbackConstants.java
